Update pico_editor.php
do_open(), do_save(), and do_delete() now support subdirectories and index.md files. Instead of using basename() on the URL, we are now using parse_url().

Currently, Pico does not support files with the same names as directories. This plugin will use the correct version based on a trailing slash being used for directories and no trailing slash being used for files. However, this will not fix that the directory will override the file on the Pico front end.